Rework client management
Creating a client using artifex_rpc is done using ArtifexClient::connect(), i.e.
the client is created on connection.

However, the application currently creates a client, which can connect or
disconnect at will.

So rework the code to be compliant with ArtifexClient, by moving the ownership
of the client to the application window.

Disconnection occurs when the client is dropped.